+++ src/BizHawk.Emulation.Cores/Consoles/Nintendo/GBHawk/Mappers/Mapper_MBC3.cs
@@ -69,7 +69,9 @@ namespace BizHawk.Emulation.Cores.Nintendo.GBHawk
 					}
 					else
 					{
-						return Core.cpu.TotalExecutedCycles > (Core.bus_access_time + 8) ? 0xFF : Core.bus_value;
+						return Core.cpu.TotalExecutedCycles > (Core.bus_access_time + 8)
+							? (byte) 0xFF
+							: Core.bus_value;
 					}
 				}
 
@@ -85,7 +87,9 @@ namespace BizHawk.Emulation.Cores.Nintendo.GBHawk
 			}
 			else
 			{
-				return Core.cpu.TotalExecutedCycles > (Core.bus_access_time + 8) ? 0xFF : Core.bus_value;
+				return Core.cpu.TotalExecutedCycles > (Core.bus_access_time + 8)
+					? (byte) 0xFF
+					: Core.bus_value;
 			}
 		}
